// src/app/services/openapi/model/organism.ts
/**
 * Digicatch
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tag } from './tag';


export interface Organism { 
    name: string;
    description: string;
    image: string;
    tags: Array<Tag>;
    readonly deleted: boolean;
    readonly updatedAt: string;
}