// src/app/services/openapi/model/catchZone.ts
/**
 * Digicatch
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Area } from './area';
import { Organism } from './organism';


export interface CatchZone { 
    name: string;
    lon: number;
    lat: number;
    radius: number;
    organisms: Array<Organism>;
    area: Area;
    readonly deleted: boolean;
    readonly updatedAt: string;
}